+++ e2e/tests-dfx/start.bash
@@ -228,6 +228,55 @@ teardown() {
     assert_match "dfx is already running"
 }
 
+@test "dfx start for shared network warns about default settings specified in dfx.json that do not apply" {
+    dfx_new hello
+
+    IGNORED_MESSAGE="Ignoring the 'defaults' field in dfx.json because project settings never apply to the shared network."
+    APPLY_SETTINGS_MESSAGE="To apply these settings to the shared network, define them in .*/config-root/.config/dfx/networks.json like so"
+
+    jq 'del(.defaults)' dfx.json | sponge dfx.json
+    jq '.defaults.bitcoin.enabled=true' dfx.json | sponge dfx.json
+    assert_command dfx start --background
+    assert_contains "$IGNORED_MESSAGE"
+    assert_match "$APPLY_SETTINGS_MESSAGE"
+    assert_contains '"bitcoin": {'
+    assert_not_contains '"replica"'
+    assert_not_contains '"canister_http"'
+    assert_command dfx stop
+
+    jq 'del(.defaults)' dfx.json | sponge dfx.json
+    jq '.defaults.replica.log_level="info"' dfx.json | sponge dfx.json
+    assert_command dfx start --background
+    assert_contains "$IGNORED_MESSAGE"
+    assert_match "$APPLY_SETTINGS_MESSAGE"
+    assert_not_contains '"bitcoin"'
+    assert_contains '"replica": {'
+    assert_not_contains '"canister_http"'
+    assert_command dfx stop
+
+    jq 'del(.defaults)' dfx.json | sponge dfx.json
+    jq '.defaults.canister_http.enabled=false' dfx.json | sponge dfx.json
+    assert_command dfx start --background
+    assert_contains "$IGNORED_MESSAGE"
+    assert_match "$APPLY_SETTINGS_MESSAGE"
+    assert_not_contains '"bitcoin"'
+    assert_not_contains '"replica"'
+    assert_contains '"canister_http": {'
+    assert_command dfx stop
+
+    jq 'del(.defaults)' dfx.json | sponge dfx.json
+    jq '.defaults.bitcoin.enabled=true' dfx.json | sponge dfx.json
+    jq '.defaults.replica.log_level="info"' dfx.json | sponge dfx.json
+    jq '.defaults.canister_http.enabled=false' dfx.json | sponge dfx.json
+    assert_command dfx start --background
+    assert_contains "$IGNORED_MESSAGE"
+    assert_match "$APPLY_SETTINGS_MESSAGE"
+    assert_contains '"bitcoin": {'
+    assert_contains '"replica": {'
+    assert_contains '"canister_http": {'
+    assert_command dfx stop
+}
+
 @test "dfx starts replica with correct log level - project defaults" {
     dfx_new
     jq '.defaults.replica.log_level="warning"' dfx.json | sponge dfx.json